requirements.txt
pillow==11.1.0
py2app==0.28.8